Master:

|       Method | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |   False |   609.1 ns |  2.52 ns |  2.11 ns |         - |        - |
| ContinueWith |   False |   905.7 ns |  4.00 ns |  3.13 ns |         - |        - |
|   AsyncAwait |    True | 2,943.7 ns | 26.80 ns | 22.38 ns |         - |    712 B |
| ContinueWith |    True | 3,257.1 ns | 45.46 ns | 42.53 ns |         - |    416 B |

This PR:

|       Method | Pending |       Mean |    Error |   StdDev | Allocated | Survived |
|------------- |-------- |-----------:|---------:|---------:|----------:|---------:|
|   AsyncAwait |   False |   560.0 ns |  5.10 ns |  4.52 ns |         - |        - |
| ContinueWith |   False |   826.1 ns |  2.62 ns |  2.32 ns |         - |        - |
|   AsyncAwait |    True | 2,691.4 ns | 18.01 ns | 14.06 ns |         - |    688 B |
| ContinueWith |    True | 3,054.0 ns | 23.37 ns | 20.72 ns |         - |    416 B |